Reduce test bloat in proxy-related tests
